+++ shader-browser/src/App.svelte
@@ -1,5 +1,0 @@
-<script lang="ts">
-  import ShaderBrowser from './lib/components/ShaderBrowser.svelte';
-</script>
-
-<ShaderBrowser />